Added skeleton table loader
