+++ apps/scoutgame/components/claim/components/BuilderRewardsScreen/BuilderRewardsScreen.tsx
@@ -1,0 +1,62 @@
+import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
+import ChevronRightIcon from '@mui/icons-material/ChevronRight';
+import { IconButton, Stack, Typography } from '@mui/material';
+import {
+  getLastWeek,
+  currentSeason,
+  getPreviousWeek,
+  getNextWeek,
+  currentSeasonNumber,
+  getSeasonWeekFromISOWeek
+} from '@packages/scoutgame/dates';
+import Link from 'next/link';
+import { Suspense } from 'react';
+
+import { TabsMenu } from 'components/common/Tabs/TabsMenu';
+
+import { LoadingTable } from '../common/LoadingTable';
+
+import { BuilderRewardsTable } from './BuilderRewardsTable';
+
+export function BuilderRewardsScreen({ period }: { period: string }) {
+  const isSeason = period === 'season';
+  const lastWeek = getLastWeek();
+  const week = isSeason ? null : period || lastWeek;
+  const previousWeek = week ? (week === currentSeason ? null : getPreviousWeek(week)) : null;
+  const nextWeek = week ? (week === lastWeek ? null : getNextWeek(week)) : null;
+
+  return (
+    <Stack gap={1} alignItems='center'>
+      <Typography color='secondary' variant='h6'>
+        Builder Rewards
+      </Typography>
+      <TabsMenu
+        tabs={[
+          { value: week || lastWeek, label: 'Weekly' },
+          { value: 'season', label: 'Season Total' }
+        ]}
+        value={isSeason ? 'season' : week || lastWeek}
+      />
+      <Stack direction='row' gap={1} alignItems='center'>
+        <Link href={previousWeek ? `/claim?tab=${previousWeek}` : ''}>
+          <IconButton disabled={!previousWeek} size='small'>
+            <ChevronLeftIcon />
+          </IconButton>
+        </Link>
+        <Typography>
+          {!week
+            ? `Season ${currentSeasonNumber}`
+            : `Week ${getSeasonWeekFromISOWeek({ season: currentSeason, week })}`}
+        </Typography>
+        <Link href={nextWeek ? `/claim?tab=${nextWeek}` : ''}>
+          <IconButton disabled={!nextWeek} size='small'>
+            <ChevronRightIcon />
+          </IconButton>
+        </Link>
+      </Stack>
+      <Suspense fallback={<LoadingTable />}>
+        <BuilderRewardsTable week={week} />
+      </Suspense>
+    </Stack>
+  );
+}
